+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Upgrade Gradle Wrapper to `8.8`
 - Dependencies - upgrade `org.jetbrains.kotlinx.kover` to `0.8.1`
 - Removed the confirmation of disabling a Ruby LSP feature
+- Added support for resetting Ruby LSP features to default (https://github.com/vitallium/intellij-plugin-ruby-lsp/issues/20)
 
 ### Deprecated